+++ packages/expo-camera/src/CameraView.tsx
@@ -115,6 +115,7 @@ export default class CameraView extends Component<CameraViewProps> {
    * Get picture sizes that are supported by the device.
    * @return Returns a Promise that resolves to an array of strings representing picture sizes that can be passed to `pictureSize` prop.
    * The list varies across Android devices but is the same for every iOS.
+   *
    */
   async getAvailablePictureSizesAsync(): Promise<string[]> {
     return (await this._cameraRef.current?.getAvailablePictureSizes()) ?? [];
@@ -276,6 +277,8 @@ export default class CameraView extends Component<CameraViewProps> {
 
   /**
    * Stops recording if any is in progress.
+   * @platform android
+   * @platform ios
    */
   stopRecording() {
     this._cameraRef.current?.stopRecording();