+++ packages/expo-camera/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### 💡 Others
 
-- [Web] Remove `forwardRef` from Camera component.
+- [Web] Remove `forwardRef` from Camera component. ([#36320](https://github.com/expo/expo/pull/36320) by [@alanjhughes](https://github.com/alanjhughes))
 
 ## 16.1.3 — 2025-04-21